requirements.txt
requests==2.31.0
python-dotenv==1.0.0
Mastodon.py==1.8.0
Pillow==10.1.0
apscheduler==3.10.4
